requirements.txt
aiocontextvars==0.2.2
aiohttp==3.5.4
aiomysql==0.0.20
aiopg==1.0.0
aioredis==1.2.0
alembic==1.0.11
amqp==2.5.0
APScheduler==3.6.1
arrow==0.14.2
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
bcrypt==3.1.7
billiard==3.6.0.0
bson==0.5.8
cached-property==1.5.1
celery==4.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
contextvars==2.4
cryptography==2.7
cssselect==1.1.0
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
imageio==2.5.0
immutables==0.11
importlib-metadata==0.20
incremental==17.5.0
jsonschema==3.0.1
kombu==4.6.3
loguru==0.4.0
lxml==4.4.1
Mako==1.0.14
MarkupSafe==1.1.1
mock==3.0.5
more-itertools==7.2.0
multidict==4.5.2
parsel==1.5.2
peewee==3.13.1
peewee-async==0.7.0
peewee-migrate==1.1.6
Pillow==8.1.0
psycopg2-binary==2.8.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.8.2
PyJWT==1.7.1
PyMySQL==0.9.2
pyOpenSSL==19.0.0
pypinyin==0.35.4
pyrsistent==0.15.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1.1
qcloudsms-py==0.1.3
queuelib==1.5.0
redis==3.2.1
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
Scrapy==1.7.4
service-identity==18.1.0
six==1.12.0
tornado==6.0.3
Twisted==19.7.0
typing-extensions==3.7.4
tzlocal==2.0.0
urllib3==1.25.7
vine==1.3.0
w3lib==1.21.0
win32-setctime==1.0.1
yarl==1.3.0
zipp==0.6.0
zope.interface==4.6.0
